session5_pose_est/requirements.txt
onnxruntime==1.4.0
Pillow==7.2.0
opencv-python-headless
requests_toolbelt
